Update trait to be as specified in OAM spec (#504)
* Updating the traits input to adhere to strict schema as specified in OAM spec.

Adding version for the traits in the name, so that it enable updating/rolling different version of the trait. (We can discuss introducing a new field "type" which could be the right way to do it).

For backward compatability, I have kept the exiting trait name as it is. But after next milestone we could delete them.

* remove ignored files that got committed

* Updating after clippy run.

* Fix Indentation.