Addressing issue# 514. (#515)
https://github.com/oam-dev/rudr/issues/514

The issue was that rudr was not reconciling it's state with the latest status.
If a deployment/svc/ing is deleted when the application confgiuration still exists, runtime doesn't recreate them. It detects those objects are not found but no action was taken.

The fix here is to recreate those objects back. The recreated objects have correct owner reference so no objects are orphaned.